pyproject.toml
[project]
name = "hemiplon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.76.0",
    "pdf2image>=1.17.0",
    "pillow>=11.2.1",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
